+++ src/main/java/dto/string_calculator/Number.java
@@ -1,3 +1,5 @@
+package dto.string_calculator;
+
 public class Number {
     private final long number;
