+++ src/main/java/dto/string_calculator/Operation.java
@@ -1,3 +1,5 @@
+package dto.string_calculator;
+
 public class Operation {
     private final String operation;
