+++ src/main/java/dto/string_calculator/Operations.java
@@ -1,3 +1,5 @@
+package dto.string_calculator;
+
 import java.util.List;
 
 public class Operations {
