+++ src/main/java/dto/string_calculator/TotalNumber.java
@@ -1,3 +1,5 @@
+package dto.string_calculator;
+
 public class TotalNumber {
     private long total;
