--- src/main/java/vo/LotteryTickets.java
@@ -1,15 +1,0 @@
-package vo;
-
-import java.util.List;
-
-public class LotteryTickets {
-    private final List<LotteryTicket> lotteryTickets;
-
-    public LotteryTickets(List<LotteryTicket> lotteryTickets) {
-        this.lotteryTickets = lotteryTickets;
-    }
-
-    public List<LotteryTicket> getLotteryTickets() {
-        return lotteryTickets;
-    }
-}
